Refactor the getAllFiles function
- move the function to separate file
- update table tool to wait until drillable item available

JIRA: RAIL-4023
